Rust API: Switch from f64 to f32 throughout protocol and client
The wire format maxes out at f32 or scaled integers, so f64 provided
no precision benefit. Using f32 halves the size of command/query
structs and is more appropriate for the no_std protocol crate on
embedded platforms where f64 may require software emulation.

+++ lib/rust/examples/bandwidth_test.rs
@@ -110,7 +110,7 @@ async fn main() -> Result<(), moteus::Error> {
             .iter()
             .map(|c| {
                 Request::new(c.make_position_command(
-                    &PositionCommand::new().position(f64::NAN).velocity(0.0),
+                    &PositionCommand::new().position(f32::NAN).velocity(0.0),
                     true,
                 ).into_frame())
             })
